examples/firecrawl_shopping_scraper/requirements.txt
upsonic
firecrawl-py
python-dotenv
anthropic
